backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
langchain==0.2.1
langchain-community==0.2.1
pydantic==2.7.1
google-generativeai==0.5.4
chromadb==0.5.3
pypdf==4.2.0
rank_bm25==0.2.2
nltk==3.8.1
python-dotenv==1.0.1
sentence-transformers==2.2.2
torch>=2.2.0
transformers>=4.35.0
accelerate>=0.30.0
huggingface_hub==0.23.0  # <--- ADD THIS LINE